Backend/requirements.txt
blinker==1.6.2
bson==0.5.10
click==8.1.6
colorama==0.4.6
dnspython==2.4.1
Flask==2.3.2
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
MarkupSafe==2.1.3
nltk==3.8.1
numpy==1.25.2
pandas==2.0.3
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.6.3
scikit-learn==1.3.0
scikit-multilearn==0.2.0
scipy==1.11.1
six==1.16.0
threadpoolctl==3.2.0
tqdm==4.65.0
tzdata==2023.3
Werkzeug==2.3.6
